+++ untitled/test.py
@@ -1,0 +1,18 @@
+from flask import Flask
+from flask.ext.sqlalchemy import SQLAlchemy
+
+app = flask(__name__)
+app.config['SWLALCHEMY_DATABASE_URI'] = 'sqlite:test.db'
+db = SQLAlchemy(app)
+
+class User(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(80), unique=True)
+    email = db.Column(db.String(120), unique=True)
+
+    def __init__(self, username, email):
+        self.username = username
+        self.email = email
+
+    def __repr__(self):
+        return '<User %r>' % self.username
